Added functionality for serial port to Avoid carashing
Fixed BUG where serialport would crash because it would not find the USB COM Port it was bound to when the server was started, so insted we look for an Arduino USB Cord in the list of COMS, then we connect to that if it exist otherwise, we ignore.

Futhermore, i added functionality for the Node App to listen to USB Connections, so that if an Arduino Cable was Released or Connected, the App would find it and Connect to it seemlesly.